+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * Changed control to switch modes to Q (dropping item).
   * Previously Shift + Right-click.
+* `/eas give` selects the tool if you already have it in your inventory.